Split API to provide explicit memory allocation
The API is now layered into two functions:
`t2s_convert`
`t2s_convert_noalloc`

The _noalloc version contains all of the actual logic, and the original one is a wrapper that does allocation.

The way we handle explicit allocation, is by running the _noalloc function twice.
It will fill out a struct telling us how much memory to give it.
Then the calling code fills in the pointers in the struct, and re-calls the function.
The second time it will actually run.

We have a very simple arena/region allocator, just bumping the pointer along.
Right now we only have two texture buffers so this doesn't do a lot of heavy lifting.
We may actually want to remove this if it feels simpler.